changing name and release number/date
